[tb/LEMS-2758/expression-public-options] Move util files to perseus-core and rename
